Fix minor issues and refactoring

#20
